+++ backend/SmartPlaylist/Getters/GenreGetter.cs
@@ -9,8 +9,10 @@ namespace SmartPlaylist.Domain
     {
         public static string[] Get()
         {
-            return Plugin.Instance.LibraryManager.GetGenres(new InternalItemsQuery()).Items.Select(x => x.Item1.ToString())
+            string[] genres = Plugin.Instance.LibraryManager.GetGenres(new InternalItemsQuery()).Items.Select(x => x.Item1.ToString())
                 .Distinct().OrderBy(x => x).ToArray();
+
+            return genres.Length == 0 ? new string[] { "None" } : genres;
         }
     }
 }