+++ backend/SmartPlaylist/Plugin.cs
@@ -94,20 +94,20 @@ namespace SmartPlaylist
             {
                 new PluginPageInfo
                 {
-                    Name = "smartplaylist.2.1.0.5.html",
-                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.1.0.5.html",
+                    Name = "smartplaylist.2.2.0.1.html",
+                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.2.0.1.html",
                     EnableInMainMenu = true,
                     MenuIcon = "subscriptions"
                 },
                 new PluginPageInfo
                 {
-                    Name = "smartplaylist.2.1.0.5.css",
-                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.1.0.5.css"
+                    Name = "smartplaylist.2.2.0.1.css",
+                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.2.0.1.css"
                 },
                 new PluginPageInfo
                 {
-                    Name = "smartplaylist.2.1.0.5.js",
-                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.1.0.5.js"
+                    Name = "smartplaylist.2.2.0.1.js",
+                    EmbeddedResourcePath = GetType().Namespace + ".Configuration.smartplaylist.2.2.0.1.js"
                 }
 
             };